requirements.txt
# requirements-docker.txt
--extra-index-url https://download.pytorch.org/whl/cpu

python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
transformers==4.36.2
sentencepiece==0.1.99
accelerate==0.25.0
mlflow==2.9.2
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6
prometheus-client==0.19.0
psutil==5.9.7
requests==2.31.0
tqdm==4.66.1
click==8.1.7
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
